+++ migrations/0018_slow_dark_phoenix.sql
@@ -1,0 +1,4 @@
+ALTER TABLE "accounts" ADD COLUMN "is_demo" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "tracks" ADD COLUMN "demo_seed" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "is_demo" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "demo_expires_at" timestamp with time zone;